% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agedecision.R
\docType{package}
\name{agedecision}
\alias{agedecision}
\alias{agedecision-package}
\title{Bilaga : Beslutsanalys av åldersbedömningar inom asylprocessen}
\description{
Bilaga : Beslutsanalys av åldersbedömningar inom asylprocessen
}
