% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{classified_adults}
\alias{classified_adults}
\title{Andel som klassificeras som vuxna.}
\usage{
classified_adults(by = 0.1, age.min = 15, age.max = 21)
}
\arguments{
\item{by}{Steglängd för prevalensen mellan 0 och 1, default är 0.1}

\item{age.min}{Lägsta ålder i åldersfördelningen, se ekvation (1).}

\item{age.max}{Högsta ålder i åldersfördelningen, se ekvation (1).}
}
\description{
Andel som klassificeras som vuxna mot prevalensen för alla alternativ/metoder och även specificitet, sensitivet se ekvation (15) och accuracy.
}
