% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunif2.R
\name{dunif2}
\alias{dunif2}
\title{Populationsantagande, två likformiga fördelningar}
\usage{
dunif2(x, min = 15, max = 21, prevalence = 0.84)
}
\arguments{
\item{x}{age.}

\item{min}{Lägsta ålder för barn, måste vara 18 år eller under. Förvalt är 15 år.}

\item{max}{Högsta ålder, måste vara 18 år eller över. Förvalt är 21 år.}

\item{prevalence}{Prevalens, andelen som är vuxna (18 år eller över). Förvalt är 0.84}
}
\description{
Två likformiga fördelningar ihopsatta till en, ena delen barn (under 18 år) och andra delen vuxna (18 år eller över). Ger täthetsfunktionen. Se Figur 2.
}
