% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{eu_all_comb}
\alias{eu_all_comb}
\title{Förväntad nytta för alla alternativ/metoder för alla kombinationer.}
\usage{
eu_all_comb(by = 0.1)
}
\arguments{
\item{by}{Minsta steg mellan 0 och 1.}
}
\description{
Förväntad nytta (Expected Utility) för alla metoder och kombinationer av prevalens och nytta för en felklassificerad vuxen, se ekvation (16) och (17).
}
