% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{eu_i}
\alias{eu_i}
\title{Förväntad nytta}
\usage{
eu_i(pM, prevalence = 0.84, lowest.utility = 0, age.min = 15,
  age.max = 21, zero.limit.child = 15, zero.limit.adult = 21)
}
\arguments{
\item{pM}{Sannolikhetsfunktion.}

\item{prevalence}{Prevalens, andel vuxna.}

\item{lowest.utility}{Nyttan för en felklassificerad vuxen.}

\item{age.min}{Lägsta ålder i åldersfördelningen, se ekvation (1).}

\item{age.max}{Högsta ålder i åldersfördelningen, se ekvation (1).}

\item{zero.limit.child}{Anger ålder då lägsta nytta för en felaktig klassificering av ett barn, se ekvation (3) och (4).}

\item{zero.limit.adult}{Anger ålder då lägsta nytta för en felklassificering av en vuxen, se ekvation (3) och (4).}
}
\description{
Förväntad nytta (Expected Utility), se ekvation (16) och (17).
}
