% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hand}
\alias{hand}
\title{Data för handled}
\format{data.frame med 3 kolumner och 6 rader.
\describe{
  \item{age}{Faktisk ålder}
  \item{mean_ba}{Medelvärde för justerad benålder givet faktisk ålder.}
  \item{sd_ba}{Standardavvikelsen för justerad benålder givet faktiskt ålder.}...
}}
\usage{
hand
}
\description{
Data för handled
}
\references{
Thodberg m. fl. 2017
}
\keyword{datasets}
