% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kna}
\alias{kna}
\title{Data för knäled}
\format{Dataset med 5 kolumner och 14 rader.
\describe{
  \item{age}{Faktisk ålder}
  \item{n}{Antal studiedeltagare givet ålder.}
  \item{moget}{Antal studiedeltagare med mogen knäled.}
  \item{omoget}{Antal studiedeltagare med omogen knäled.}
}}
\usage{
kna
}
\description{
Data för knäled
}
\references{
Krämer m. fl. (2014)

Saint-Martin m. fl. (2015)
}
\keyword{datasets}
