% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{logcoef}
\alias{logcoef}
\title{Logistiska regressionskoefficienter}
\format{Dataset med 5 kolumner och 14 rader.
\describe{
  \item{intercept}{}
  \item{slope}{}
  \item{method}{Vilken metod som avses.}
}}
\usage{
logcoef
}
\description{
De logistiska regressionskoefficienterna som skattats från visdomstand (tand),
knäled (knä) och handled (hand).
}
\keyword{datasets}
