% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{logistic_center}
\alias{logistic_center}
\title{Centrerad ålder för en logistisk funktion}
\usage{
logistic_center(intercept, slope)
}
\arguments{
\item{intercept}{Interceptet från logistisk regression.}

\item{slope}{Lutningen från logistisk regression.}
}
\description{
Givet regressionskoefficienterna från en logistisk regression
erhålls testets centrerade ålder, vilket innebär den åldern där 50% bedöms med
mogen respektive omogen kroppsdel.
}
