% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{logistic_function}
\alias{logistic_function}
\title{Logistisk funktion}
\usage{
logistic_function(x, intercept, slope)
}
\arguments{
\item{x}{Ålder.}

\item{intercept}{Interceptet från en logistisk regression.}

\item{slope}{Lutningen från en logistisk regression.}
}
\description{
Logistisk funktion som ger sannolikheten att kroppsdelen bedöms mogen givet ålder, se ekvation (7).
}
