% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{pM}
\alias{pM}
\alias{pM_tand}
\alias{pM_kna}
\alias{pM_hand_18}
\alias{pM_hand_19}
\alias{pM_hand_20}
\alias{pM_RMV}
\alias{pM_coinflip}
\alias{pM_barn}
\alias{pM_vuxen}
\title{Sannolikhetsfunktioner från studieresultaten.}
\usage{
pM_tand(x)

pM_kna(x)

pM_hand_18(x)

pM_hand_19(x)

pM_hand_20(x)

pM_RMV(x)

pM_coinflip(x)

pM_barn(x)

pM_vuxen(x)
}
\arguments{
\item{x}{Ålder.}
}
\description{
Funktioner från studieresultaten, endast ålder (x) i år anges.
}
\seealso{
\link{logistic_function}

\link{RMV}
}
\concept{Sannolikhetsfunktioner}
