% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_functions.R
\name{RMV}
\alias{RMV}
\title{Rättsmedicinalverkets metod}
\usage{
RMV(x, intercept1, slope1, kappa1, intercept2, slope2, kappa2)
}
\arguments{
\item{intercept1}{intercept metod 1}

\item{slope1}{lutning metod 1}

\item{kappa1}{kappa metod 1}

\item{intercept2}{intercept metod 2}

\item{slope2}{lutning metod 1}

\item{kappa2}{kappa metod 1}
}
\description{
Rättsmedicinalverkets metod med två oberoende bedömare för knä respektive tand och kombinationen att minst en visar mogen.
}
