% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{spec_sens}
\alias{spec_sens}
\title{Specificitet och sensitivitet.}
\usage{
spec_sens(pM, age.min = 15, age.max = 21)
}
\arguments{
\item{pM}{Sannolikhetsfunktion.}

\item{age.min}{Lägsta ålder i åldersfördelningen, se ekvation (1).}

\item{age.max}{Högsta ålder i åldersfördelningen, se ekvation (1).}
}
\description{
Specificitet och sensitivitet, se ekvation (15).
}
