% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{spec_sens_all}
\alias{spec_sens_all}
\title{Specificitet och sensitivitet för alla alternativ/metoder.}
\usage{
spec_sens_all(age.min = 15, age.max = 21)
}
\arguments{
\item{age.min}{Lägsta ålder i åldersfördelningen, se ekvation (1).}

\item{age.max}{Högsta ålder i åldersfördelningen, se ekvation (1).}
}
\description{
Specificitet och sensitivitet för alla alternativ/metoder, se ekvation (15).
}
