% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tand}
\alias{tand}
\title{Data för visdomstand}
\format{data.frame 5 kolumner och 14 rader.
\describe{
  \item{age}{Faktisk ålder}
  \item{n}{Antal studiedeltagare givet ålder.}
  \item{moget}{Antal studiedeltagare med mogen visdomstand.}
  \item{omoget}{Antal studiedeltagare med omogen visdomstand.}
}}
\usage{
tand
}
\description{
Data för visdomstand
}
\references{
Simonsson m. fl (2017)
}
\keyword{datasets}
