% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_models.R
\name{utility_discrete_adult}
\alias{utility_discrete_adult}
\title{Diskret nyttomodell då utfallet är mogen/vuxen}
\usage{
utility_discrete_adult(x)
}
\arguments{
\item{x}{Ålder.}

\item{lowest.utility}{Den lägsta nyttan då ett barn klassificeras som ett barn. Förvalt är 0.}
}
\description{
Se ekvation 4.
}
