% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_models.R
\name{utility_discrete_child}
\alias{utility_discrete_child}
\title{Diskret nyttomodell då utfallet är omogen/barn $u_D^O(x)$}
\usage{
utility_discrete_child(x, lowest.utility = 0)
}
\arguments{
\item{x}{Ålder.}

\item{lowest.utility}{Den lägsta nyttan då em vuxen klassificeras som ett barn. Förvalt är 0.}
}
\description{
Se ekvation 4.
}
