% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_models.R
\name{utility_linear_adult}
\alias{utility_linear_adult}
\title{Linjär nyttomodell då utfallet är mogen/vuxen}
\usage{
utility_linear_adult(x, zero.limit = 15)
}
\arguments{
\item{x}{Ålder.}

\item{zero.limit}{Den högsta ålder som ger den lägsta nyttan 0. Förvalt är 15 år.}
}
\description{
Se ekvation 5.
}
