% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_models.R
\name{utility_linear_child}
\alias{utility_linear_child}
\title{Linjär nyttomodell då utfallet är omogen/barn}
\usage{
utility_linear_child(x, zero.limit = 21, lowest.utility = 0)
}
\arguments{
\item{x}{Ålder.}

\item{zero.limit}{Den lägsta ålder som ger den lägsta nyttan. Förvalt är 21 år.}

\item{lowest.utility}{Den lägsta nyttan som inträffar vid zero.limit eller högre ålder. Betecknas i uppsatsen U_21O.}
}
\description{
Se ekvation 5.
}
